bug: Check Suite id can be bigger than MAX_INT
Not sure if it can be bigger than long either, but this does not make things worse and solves the issue I ran into.

+++ src/main/java/com/spotify/github/v3/checks/CheckSuite.java
@@ -36,7 +36,7 @@ public interface CheckSuite {
    *
    * @return the integer
    */
-  Integer id();
+  Long id();
 
   Optional<App> app();
